chore: update core shared ui imports
