+++ apps/journeys-admin/pages/_app.tsx
@@ -8,7 +8,7 @@ import TagManager from 'react-gtm-module'
 import { datadogRum } from '@datadog/browser-rum'
 import { CacheProvider } from '@emotion/react'
 import type { EmotionCache } from '@emotion/cache'
-import { createEmotionCache } from '@core/shared/ui'
+import { createEmotionCache } from '@core/shared/ui/createEmotionCache'
 import { appWithTranslation } from 'next-i18next'
 import { useTranslation } from 'react-i18next'
 import { useApollo } from '../src/libs/apolloClient'
