+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Card/BackgroundColor/BackgroundColor.tsx
@@ -12,7 +12,8 @@ import ColorizeIcon from '@mui/icons-material/Colorize'
 import { HexColorPicker } from 'react-colorful'
 import { gql, useMutation } from '@apollo/client'
 import { useEditor, TreeBlock, useJourney } from '@core/journeys/ui'
-import { TabPanel, tabA11yProps, themes } from '@core/shared/ui'
+import { TabPanel, tabA11yProps } from '@core/shared/ui/TabPanel'
+import { themes } from '@core/shared/ui/themes'
 import { CardFields } from '../../../../../../../../__generated__/CardFields'
 import { CardBlockBackgroundColorUpdate } from '../../../../../../../../__generated__/CardBlockBackgroundColorUpdate'
 import { Swatch } from './Swatch'
