+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Step/NextCard/NextCard.tsx
@@ -4,7 +4,7 @@ import Divider from '@mui/material/Divider'
 import Tab from '@mui/material/Tab'
 import Tabs from '@mui/material/Tabs'
 import { useTheme } from '@mui/material/styles'
-import { TabPanel, tabA11yProps } from '@core/shared/ui'
+import { TabPanel, tabA11yProps } from '@core/shared/ui/TabPanel'
 import { Conditions } from './Conditions'
 import { Cards } from './Cards'
 import { SelectedCard } from './SelectedCard'
