+++ apps/journeys-admin/src/components/Editor/ControlPanel/ColorDisplayIcon/ColorDisplayIcon.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react'
 import Paper from '@mui/material/Paper'
 import Box from '@mui/material/Box'
-import { ThemeProvider } from '@core/shared/ui'
+import { ThemeProvider } from '@core/shared/ui/ThemeProvider'
 import { TreeBlock, useEditor, useJourney } from '@core/journeys/ui'
 import { GetJourney_journey_blocks_CardBlock as CardBlock } from '../../../../../__generated__/GetJourney'
 import {
