+++ apps/journeys-admin/src/components/Editor/ControlPanel/ControlPanel.tsx
@@ -3,7 +3,7 @@ import Tab from '@mui/material/Tab'
 import Tabs from '@mui/material/Tabs'
 import { ReactElement, SyntheticEvent } from 'react'
 import { TreeBlock, useEditor, ActiveTab, ActiveFab } from '@core/journeys/ui'
-import { TabPanel, tabA11yProps } from '@core/shared/ui'
+import { TabPanel, tabA11yProps } from '@core/shared/ui/TabPanel'
 import { GetJourney_journey_blocks_StepBlock as StepBlock } from '../../../../__generated__/GetJourney'
 import { CardPreview } from '../../CardPreview'
 import { Attributes } from './Attributes'
