+++ apps/journeys-admin/src/components/JourneyList/JourneySort/JourneySort.tsx
@@ -16,7 +16,7 @@ import FormControlLabel from '@mui/material/FormControlLabel'
 import Popover from '@mui/material/Popover'
 import RadioGroup from '@mui/material/RadioGroup'
 import Radio from '@mui/material/Radio'
-import { useBreakpoints } from '@core/shared/ui'
+import { useBreakpoints } from '@core/shared/ui/useBreakpoints'
 
 export enum SortOrder {
   CREATED_AT = 'createdAt',
