+++ apps/journeys-admin/src/components/JourneyList/StatusTabPanel/StatusTabPanel.tsx
@@ -1,7 +1,7 @@
 import { ReactElement, SyntheticEvent, useState } from 'react'
 import { sortBy } from 'lodash'
 import Card from '@mui/material/Card'
-import { TabPanel, tabA11yProps } from '@core/shared/ui'
+import { TabPanel, tabA11yProps } from '@core/shared/ui/TabPanel'
 import { Theme } from '@mui/material/styles'
 import Tab from '@mui/material/Tab'
 import Tabs from '@mui/material/Tabs'
