+++ apps/journeys-admin/src/components/JourneyView/Properties/Properties.tsx
@@ -6,7 +6,7 @@ import Drawer from '@mui/material/Drawer'
 import Toolbar from '@mui/material/Toolbar'
 import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
-import { CopyTextField } from '@core/shared/ui'
+import { CopyTextField } from '@core/shared/ui/CopyTextField'
 import { useJourney } from '@core/journeys/ui'
 import { useTranslation } from 'react-i18next'
 import EditIcon from '@mui/icons-material/Edit'
