+++ apps/journeys/pages/_app.tsx
@@ -7,7 +7,7 @@ import TagManager from 'react-gtm-module'
 import { datadogRum } from '@datadog/browser-rum'
 import { CacheProvider } from '@emotion/react'
 import type { EmotionCache } from '@emotion/cache'
-import { createEmotionCache } from '@core/shared/ui'
+import { createEmotionCache } from '@core/shared/ui/createEmotionCache'
 import { SnackbarProvider } from 'notistack'
 import { getAuth, onAuthStateChanged } from 'firebase/auth'
 import { appWithTranslation } from 'next-i18next'
