+++ apps/journeys/pages/index.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from 'react'
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
 import Container from '@mui/material/Container'
-import { ThemeProvider } from '@core/shared/ui'
+import { ThemeProvider } from '@core/shared/ui/ThemeProvider'
 import Link from 'next/link'
 import { GetStaticProps } from 'next'
 import { gql } from '@apollo/client'
