+++ apps/journeys/src/components/Conductor/Conductor.spec.tsx
@@ -1,4 +1,4 @@
-import { useBreakpoints } from '@core/shared/ui'
+import { useBreakpoints } from '@core/shared/ui/useBreakpoints'
 import {
   activeBlockVar,
   JourneyProvider,
