+++ apps/watch-admin/pages/_app.tsx
@@ -6,7 +6,7 @@ import { SnackbarProvider } from 'notistack'
 import { DefaultSeo } from 'next-seo'
 import { CacheProvider } from '@emotion/react'
 import type { EmotionCache } from '@emotion/cache'
-import { createEmotionCache } from '@core/shared/ui'
+import { createEmotionCache } from '@core/shared/ui/createEmotionCache'
 import { useApollo } from '../src/libs/apolloClient'
 import { ThemeProvider } from '../src/components/ThemeProvider'
 import { initAuth } from '../src/libs/firebaseClient/initAuth'
