+++ apps/watch/src/components/Videos/VideoList/List/VideoListList.tsx
@@ -6,7 +6,7 @@ import ListItemText from '@mui/material/ListItemText'
 import Typography from '@mui/material/Typography'
 import LoadingButton from '@mui/lab/LoadingButton'
 import { ReactElement } from 'react'
-import { secondsToTimeFormat } from '@core/shared/ui'
+import { secondsToTimeFormat } from '@core/shared/ui/timeFormat'
 import AddRounded from '@mui/icons-material/AddRounded'
 import ListItem from '@mui/material/ListItem'
 import Link from 'next/link'
