+++ apps/watch/src/libs/storybook/config.tsx
@@ -1,4 +1,4 @@
-import { sharedUiConfig } from '@core/shared/ui'
+import { sharedUiConfig } from '@core/shared/ui/sharedUiConfig'
 
 // Must set parameters at component level for shared-storybook stories to work
 export const watchConfig = {
