+++ libs/journeys/ui/src/components/Card/Card.spec.tsx
@@ -1,5 +1,5 @@
 import { render, waitFor } from '@testing-library/react'
-import { themes } from '@core/shared/ui'
+import { themes } from '@core/shared/ui/themes'
 import { TreeBlock, blurImage } from '../..'
 import { ThemeMode, ThemeName } from '../../../__generated__/globalTypes'
 import { ImageFields } from '../Image/__generated__/ImageFields'
