+++ libs/journeys/ui/src/libs/storybook/containers.tsx
@@ -1,6 +1,6 @@
 import Paper from '@mui/material/Paper'
 import { ReactElement, ReactNode } from 'react'
-import { ThemeProvider } from '@core/shared/ui'
+import { ThemeProvider } from '@core/shared/ui/ThemeProvider'
 import { ThemeMode, ThemeName } from '../../../__generated__/globalTypes'
 
 interface StoryCardProps {
