--- libs/shared/ui/src/index.ts
@@ -1,13 +1,0 @@
-export { CopyTextField } from './components/CopyTextField'
-export { NextImage } from './components/NextImage'
-export { ThemeProvider } from './components/ThemeProvider'
-export { TabPanel, tabA11yProps } from './components/TabPanel'
-export { sharedUiConfig, simpleComponentConfig } from './libs/storybook/config'
-export { themes } from './libs/themes'
-export { useBreakpoints } from './libs/useBreakpoints'
-export { createEmotionCache } from './libs/emotion/cache'
-export {
-  secondsToTimeFormat,
-  timeFormatToSeconds,
-  secondsToMinutes
-} from './libs/timeFormat'
