+++ libs/shared/ui/src/libs/ThemeDecorator/index.ts
@@ -1,0 +1,1 @@
+export { ThemeDecorator } from './ThemeDecorator'
