+++ libs/shared/ui/src/libs/createEmotionCache/index.ts
@@ -1,0 +1,1 @@
+export { createEmotionCache } from './createEmotionCache'
