+++ libs/shared/ui/src/libs/sharedUiConfig/index.ts
@@ -1,0 +1,1 @@
+export { sharedUiConfig } from './sharedUiConfig'
