+++ libs/shared/ui/src/libs/simpleComponentConfig/index.ts
@@ -1,0 +1,1 @@
+export { simpleComponentConfig } from './simpleComponentConfig'
