+++ libs/shared/ui/src/libs/simpleComponentConfig/simpleComponentConfig.tsx
@@ -1,0 +1,13 @@
+import { sharedUiConfig } from '../sharedUiConfig'
+
+// Simple components are not responsive, simplify VR testing
+export const simpleComponentConfig = {
+  ...sharedUiConfig,
+  parameters: {
+    ...sharedUiConfig.parameters,
+    chromatic: {
+      ...sharedUiConfig.parameters?.chromatic,
+      viewports: [600]
+    }
+  }
+}
